% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Compute pair-wise correlations and hypothesis test}
\usage{
corr_matrix(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
list with two tables: corrs (correlations), pvals (p-values)
}
\description{
Computes pair-wise correlations between variables in a dataframe df
Uses p-values to test:\cr\cr
H0: rho = 0\cr
H1: rho != 0
}
\examples{
library(sda123)
corr_matrix(mtcars[,c("mpg","hp","drat","wt")])
}
