% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ericsson}
\alias{ericsson}
\title{Daily percentage returns on Ericsson B stock}
\format{
A data frame with 25 rows and 2 variables:
\describe{
\item{datum}{date in format YYYY-MM-DD}
\item{avkastning}{daily percentage returns 100*(log(x_t)-log(x_{t-1})}
...
}
}
\source{
Nasdaq Nordic \url{https://www.nasdaqomxnordic.com/index/historiska_kurser?languageId=3&Instrument=SSE101}.
}
\usage{
ericsson
}
\description{
This data set contains daily percentage returns on Ericsson B stock for all of year 2022
}
\keyword{datasets}
