% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lifespan}
\alias{lifespan}
\title{Determinants of life expectancy in 30 countries.}
\format{
A data frame with 30 rows and 5 variables:
\describe{
\item{country}{Country name}
\item{spending}{Spending on health per capita in thousands of dollars per capita.}
\item{lifespan}{Life expectancy in years}
\item{doctorvisits}{average number of visits/consultations to the doctor}
\item{gdp}{gross domestic product per capita in thousands of dollars per capita.}
...
}
}
\source{
Gelman, Hill and Vehtari (2020). Regression and other stories, \emph{Cambridge University Press}. \url{https://avehtari.github.io/ROS-Examples/}

OECD. \url{https://data.oecd.org/}
}
\usage{
lifespan
}
\description{
Determinants of life expectancy in 30 countries.
}
\keyword{datasets}
