% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{logisticreg_summary}
\alias{logisticreg_summary}
\title{Summarize the results from a logistic regression analysis}
\usage{
logisticreg_summary(glmobject, odds_ratio = T, param = T, conf_intervals = F)
}
\arguments{
\item{glmobject}{a fitted regression model from \code{glm}.}

\item{odds_ratio}{\code{TRUE} if odds ratios for parameters is computed.}

\item{param}{\code{TRUE} if parameter estimates, standard errors etc is computed.}

\item{conf_intervals}{\code{TRUE} if confidence intervals for parameters.}
}
\value{
list with two tables: param, odds_ratio
}
\description{
Alternative to \code{summary.glm} to summarize a regression from \code{glm}.
Prints a table similar to the one generated by SAS and Minitab.
}
\examples{
library(sda123)
glmfit <- glm(survived ~ age + sex + firstclass, data = titanic, family = binomial)
logisticreg_summary(glmfit)
}
