% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{moving_average}
\alias{moving_average}
\title{Centered moving average to smooth out a time series}
\usage{
moving_average(y, r, plotfig = TRUE)
}
\arguments{
\item{y}{a vector with time series data}

\item{r}{the number of observations to the left of the center in the average,
i.e. the function computes a 2r+1 point average.}

\item{plotfig}{if TRUE then a figure is plotted with data and moving average.}
}
\value{
a vector of the same length as y with moving averages (NA at boundaries)
}
\description{
Centered moving average to smooth out a time series
}
\examples{
library(SUdatasets)
M = moving_average(globaltemp$temp, 2)
}
