% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{moving_average_manip}
\alias{moving_average_manip}
\title{Manipulate version of centered moving average to smooth out a time series}
\usage{
moving_average_manip(y)
}
\arguments{
\item{y}{a vector with time series data}
}
\description{
Manipulate version of centered moving average to smooth out a time series
}
\examples{
library(SUdatasets)
moving_average_manip(globaltemp$temp)
}
