% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{moving_average_seasonal}
\alias{moving_average_seasonal}
\title{Moving average to smooth out seasonal time series}
\usage{
moving_average_seasonal(y, season, plotfig = TRUE)
}
\arguments{
\item{y}{a vector with time series data}

\item{season}{season = 12 for montly, season = 4 for quarterly etc}

\item{plotfig}{if TRUE then a figure is plotted with data and moving average.}
}
\value{
a vector of the same length as y with moving averages (NA at boundaries)
}
\description{
Moving average to smooth out seasonal time series
}
\examples{
M = moving_average_seasonal(c(AirPassengers), season = 12)
}
