% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{poisreg_simulate}
\alias{poisreg_simulate}
\title{Simulate from a Poisson regression model with a log link}
\usage{
poisreg_simulate(
  n,
  betavect,
  intercept = TRUE,
  covdist = "normal",
  rho_x = 0,
  sigma_x = rep(1, length(betavect) - intercept)
)
}
\arguments{
\item{n}{the number of observations in the simulated dataset.}

\item{betavect}{a vector with regression coefficients
c(beta_0,beta_1,...beta_k). First element is intercept if \code{intercept = TRUE}}

\item{intercept}{if \code{TRUE} an intercept is added to the model.}

\item{covdist}{distribution of the covariates. Options: \code{'normal'} or \code{'uniform'}.}

\item{rho_x}{correlation among the covariates. Same for all covariate pairs.}

\item{sigma_x}{vector with standard deviation of the covariates.}
}
\value{
dataframe with simulated data (y, X1, X2, ..., XK) (no intercept included).
}
\description{
Simulates a dataset with \code{n} observation from the Poisson regression model \cr
\deqn{y \vert \boldsymbol{x} \sim \mathrm{Pois}(\exp(\beta_0 + \beta_1x_1 + \ldots + \beta_k x_k))}{y | **x** ~ Pois(exp(\beta_0 + \beta_1 * x_1 + ... + \beta_k * x_k))}
with covariates (\strong{x}) simulated from a normal distribution with the same correlation \code{rho_x} \cr
between all pairs of covariates. Covariate x_j has standard deviation \code{sigma_x[j]}. \cr
Alternatively the covariate can follow a uniform distribution.
}
\examples{
library(sda123)
simdata <- poisreg_simulate(n = 500, betavect = c(1, -0.2, 0.1, 0))
poisregfit <- glm(y ~ X1 + X2 + X3, data = simdata, family = poisson)
summary(poisregfit)
}
