% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{reg_crossval}
\alias{reg_crossval}
\title{K-fold cross-validation of regression models estimated with lm()}
\usage{
reg_crossval(formula, data, nfolds, obs_order = "random")
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame with the data used for fitting the models.}

\item{nfolds}{the number of folds in the cross-validation.}

\item{obs_order}{order of the observations when splitting the data. obs_order = "random" gives a random order.}
}
\value{
RMSE Root mean squared prediction error on test data
}
\description{
K-fold cross-validation of regression models estimated with lm()
}
\examples{
library(sda123)
RMSE_CV = reg_crossval(mpg ~ hp, data = mtcars, nfolds = 4, obs_order = 1:32)
print(RMSE_CV)
}
