% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{reg_predict}
\alias{reg_predict}
\title{Plot confidence and prediction intervals for simple linear regression \cr}
\usage{
reg_predict(formula, data, level = 0.95, conf_int_line = T, pred_interval = T)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a data frame with the data.}

\item{level}{confidence level, default is level = 0.95}

\item{conf_int_line}{if TRUE, then conf intervals for regression line are plotted.}

\item{pred_interval}{if TRUE, then prediction intervals are plotted.}
}
\value{
plot of data with overlayed intervals
}
\description{
Plot confidence and prediction intervals for simple linear regression \cr
}
\examples{
library(sda123)
reg_predict(mpg ~ hp, data = mtcars)
}
