% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{reg_residuals}
\alias{reg_residuals}
\title{Residual analysis mimicing the 4-in-1 plots from Minitab}
\usage{
reg_residuals(lm_object, studentized = FALSE)
}
\arguments{
\item{lm_object}{a fitted regression model from \code{lm}.}

\item{studentized}{use (externally) studentized residuals. Defaults to FALSE.}
}
\description{
Plots:
\enumerate{
\item Normal QQ-plot
\item Residuals vs fitted values
\item Histogram and normal density fit
\item Residuals vs order.
}
}
\examples{
library(sda123)
fit = lm(mpg ~ hp, data = mtcars)
reg_residuals(fit)
}
