% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{reg_summary}
\alias{reg_summary}
\title{Summarize the results from a regression analysis}
\usage{
reg_summary(
  lmobject,
  anova = T,
  fit_measures = T,
  param = T,
  conf_intervals = F,
  vif_factors = F
)
}
\arguments{
\item{lmobject}{a fitted regression model from \code{lm}.}

\item{anova}{\code{TRUE} if an ANOVA table is computed.}

\item{fit_measures}{\code{TRUE} if measures of fit (R² etc) is computed.}

\item{param}{\code{TRUE} if parameter estimates, standard errors etc is computed.}

\item{conf_intervals}{\code{TRUE} if confidence intervals for parameters.}

\item{vif_factors}{\code{TRUE} if variance inflation factors are to be printed.}
}
\value{
list with three tables: param, anova and fit_measures
}
\description{
Alternative to \code{summary.lm} to summarize a regression from \code{lm}.
Prints a table similar to the one generated by SAS and Minitab.
}
\examples{
library(sda123)
lmfit = lm(nRides ~ temp + hum + windspeed, data = bike)
regsumm = reg_summary(lmfit, anova = TRUE, conf_intervals = TRUE, vif_factors = TRUE)
regsumm$param
regsumm$anova
regsumm$fit_measures
}
