% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda123-package.R
\docType{package}
\name{sda123-package}
\alias{sda123}
\alias{sda123-package}
\title{sda123: R-paket för kurserna Statistik och dataanalys I och II, 15 hp vid SU}
\description{
Funktioner för grundläggande statistik, inkl regression.
}
\author{
\strong{Maintainer}: Mattias Villani \email{mattias.villani@gmail.com} (\href{https://orcid.org/0000-0003-2786-2519}{ORCID})

}
\keyword{internal}
