% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{simAR1}
\alias{simAR1}
\title{Simulate from an AR(1) process}
\usage{
simAR1(n, phi = 0, mu = 0, sigma_eps = 1, epsilons = NA)
}
\description{
Simulates \code{n} observations from
\deqn{x_t = \mu + \phi(x_{t-1}-\mu) + \epsilon, \epsilon \sim N(0, \sigma_\epsilon)}{x_t = \mu + \phi(x_{t-1}-\mu) + \epsilon, \epsilon ~ N(0, \sigma_ \epsilon^2)}
}
\examples{
library(sda123)
simdata = simAR1(n = 100, phi = 0.7, sigma_eps = 1)
plot(simdata)
}
