% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{triss}
\alias{triss}
\title{Winnings in the Swedish Triss lottery}
\format{
A data frame with 25 rows and 2 variables:
\describe{
\item{vinst}{amount in each winning class}
\item{antal}{number of tickets in each winning class}
\item{probs}{probability for each winning class}
...
}
}
\source{
Svenska spel \url{https://www.svenskaspel.se/triss/spelguide/triss-30}.
}
\usage{
triss
}
\description{
This data set list the number of possible winning amounts and the number of tickets in each winning class for the Swedish Triss lottery.
}
\keyword{datasets}
